+++ src/jsoncons/json.hpp
@@ -1693,6 +1693,23 @@ public:
             evaluate().erase(name);
         }
 
+        // merge
+
+        void merge(const json_type& source)
+        {
+            return evaluate().merge(source);
+        }
+
+        void merge(json_type&& source)
+        {
+            return evaluate().merge(std::forward<json_type>(source));
+        }
+
+        void merge(object_iterator hint, const json_type& source)
+        {
+            return evaluate().merge(hint, source);
+        }
+
        // set
 
         template <class T>
@@ -3312,6 +3329,56 @@ public:
         default:
             {
                 JSONCONS_THROW_EXCEPTION(std::runtime_error,"Attempting to call set on a value that is not an object");
+            }
+        }
+    }
+
+    // merge
+
+    void merge(const json_type& source)
+    {
+        switch (var_.type_id())
+        {
+        case value_type::empty_object_t:
+            create_object_implicitly();
+            // FALLTHRU
+        case value_type::object_t:
+            return object_value().merge(source.object_value());
+        default:
+            {
+                JSONCONS_THROW_EXCEPTION(std::runtime_error,"Attempting to merge a value that is not an object");
+            }
+        }
+    }
+
+    void merge(json_type&& source)
+    {
+        switch (var_.type_id())
+        {
+        case value_type::empty_object_t:
+            create_object_implicitly();
+            // FALLTHRU
+        case value_type::object_t:
+            return object_value().merge(std::move(source.object_value()));
+        default:
+            {
+                JSONCONS_THROW_EXCEPTION(std::runtime_error,"Attempting to merge a value that is not an object");
+            }
+        }
+    }
+
+    void merge(object_iterator hint, const json_type& source)
+    {
+        switch (var_.type_id())
+        {
+        case value_type::empty_object_t:
+            create_object_implicitly();
+            // FALLTHRU
+        case value_type::object_t:
+            return object_value().merge(hint, source.object_value());
+        default:
+            {
+                JSONCONS_THROW_EXCEPTION(std::runtime_error,"Attempting to merge a value that is not an object");
             }
         }
     }